pyproject.toml
[tool.poetry]
name = "release-management-examples"
version = "0.1.0"
description = ""
authors = ["Tyler Ruppert <{ID}+{username}@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
